requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.1
django-phonenumber-field==6.3.0
idna==3.3
phonenumbers==8.12.54
Pillow==9.2.0
requests==2.28.1
sqlparse==0.4.2
stripe==4.0.2
urllib3==1.26.11
